requirements.txt
Flask==3.1.3
requests==2.32.5
livekit-agents<1.0.0
livekit-plugins-openai<1.0.0
livekit-plugins-silero<1.0.0
python-dotenv==1.2.2